refactor(verifier): consolidate evidence types
