+++ packages/core/lib/v3/index.ts
@@ -84,6 +84,11 @@ export type {
   RubricCriterion,
   AgentEvidence,
   AgentEvidenceModality,
+  CanonicalEvidence,
+  CanonicalScreenshot,
+  CanonicalTextEvidence,
+  EvidenceLoadOptions,
+  EvidenceLoadResult,
   ProbeEvidence,
   ToolOutput,
   Verifier,
